Create centralized Codespaces RPC invoker
