+++ pkg/liveshare/session.go
@@ -97,32 +97,6 @@ func (s *Session) StartSSHServerWithOptions(ctx context.Context, options StartSS
 	return port, response.User, nil
 }
 
-// StartJupyterServer starts a Juypyter server in the container and returns
-// the port on which it listens and the server URL.
-func (s *Session) StartJupyterServer(ctx context.Context) (int, string, error) {
-	var response struct {
-		Result    bool   `json:"result"`
-		Message   string `json:"message"`
-		Port      string `json:"port"`
-		ServerUrl string `json:"serverUrl"`
-	}
-
-	if err := s.rpc.do(ctx, "IJupyterServerHostService.getRunningServer", []string{}, &response); err != nil {
-		return 0, "", fmt.Errorf("failed to invoke JupyterLab RPC: %w", err)
-	}
-
-	if !response.Result {
-		return 0, "", fmt.Errorf("failed to start JupyterLab: %s", response.Message)
-	}
-
-	port, err := strconv.Atoi(response.Port)
-	if err != nil {
-		return 0, "", fmt.Errorf("failed to parse JupyterLab port: %w", err)
-	}
-
-	return port, response.ServerUrl, nil
-}
-
 func (s *Session) RebuildContainer(ctx context.Context, full bool) error {
 	rpcMethod := "IEnvironmentConfigurationService.incrementalRebuildContainer"
 	if full {
